# docs/swagger.yaml
basePath: /v1/core-services
definitions:
  github_com_modami_core-service_internal_domain.CategoryOrder:
    properties:
      id:
        type: string
      sort_order:
        type: integer
    type: object
...
  internal_adapter_handler.CreateCategoryRequest:
    properties:
      icon:
        type: string
      name:
        type: string
      name_vi:
        type: string
      parent_id:
        type: string
      slug:
        type: string
      sort_order:
        type: integer
    required:
    - name
    - name_vi
    - slug
    type: object
  internal_adapter_handler.CreatePackageRequest:
    properties:
      badge_name:
        type: string
      code:
        type: string
      credits_per_month:
        type: integer
      featured_slots:
        type: integer
      name:
        type: string
      price_monthly:
        type: integer
      price_yearly:
        type: integer
      priority_support:
        type: boolean
      search_boost:
        type: boolean
      search_priority:
        type: boolean
      sort_order:
        type: integer
      tier:
        type: integer
    required:
    - code
    - name
    type: object
  internal_adapter_handler.StandardErrorEnvelope:
    properties:
      error: {}
      meta: {}
      success:
        type: boolean
    type: object
  internal_adapter_handler.StandardSuccessEnvelope:
    properties:
      data: {}
      meta: {}
      success:
        type: boolean
    type: object
  internal_adapter_handler.UpdateCategoryRequest:
    properties:
      icon:
        type: string
      name:
        type: string
      name_vi:
        type: string
      slug:
        type: string
      sort_order:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: 'Core service: catalog, orders, master data, and seller APIs for the
    Modami marketplace.'
  title: Modami Core Service API
  version: "1.0"
paths:
  /admin/categories:
    post:
      consumes:
      - application/json
      parameters:
      - description: Category
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_adapter_handler.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardSuccessEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
      security:
      - BearerAuth: []
      summary: 'Admin: create category'
      tags:
      - Admin
  /admin/categories/{id}:
    delete:
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
      security:
      - BearerAuth: []
      summary: 'Admin: delete category'
      tags:
      - Admin
    put:
      consumes:
      - application/json
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_adapter_handler.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardSuccessEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
      security:
      - BearerAuth: []
      summary: 'Admin: update category'
      tags:
      - Admin
  /admin/categories/{id}/toggle:
    put:
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardSuccessEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
      security:
      - BearerAuth: []
      summary: 'Admin: toggle category active'
      tags:
      - Admin
  /admin/categories/reorder:
    put:
      consumes:
      - application/json
      parameters:
      - description: Ordered ids
        in: body
        name: body
        required: true
        schema:
          items:
            $ref: '#/definitions/github_com_modami_core-service_internal_domain.CategoryOrder'
          type: array
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
      security:
      - BearerAuth: []
      summary: 'Admin: reorder categories'
      tags:
      - Admin
...
  /admin/packages:
    post:
      consumes:
      - application/json
      parameters:
      - description: Package
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_adapter_handler.CreatePackageRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardSuccessEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
      security:
      - BearerAuth: []
      summary: 'Admin: create package'
      tags:
      - Admin
  /admin/packages/{id}:
    put:
      consumes:
      - application/json
      parameters:
      - description: Package ID
        in: path
        name: id
        required: true
        type: string
      - description: Package fields
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_adapter_handler.CreatePackageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardSuccessEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
      security:
      - BearerAuth: []
      summary: 'Admin: update package'
      tags:
      - Admin
  /admin/packages/{id}/toggle:
    put:
      parameters:
      - description: Package ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardSuccessEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
      security:
      - BearerAuth: []
      summary: 'Admin: toggle package active'
      tags:
      - Admin
  /categories:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardSuccessEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
      summary: List categories
      tags:
      - Categories
  /categories/{slug}:
    get:
      parameters:
      - description: Category slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardSuccessEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
      summary: Get category by slug
      tags:
      - Categories
  /categories/{slug}/children:
    get:
      parameters:
      - description: Parent category slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardSuccessEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
      summary: List child categories
      tags:
      - Categories
...
  /hashtags/suggest:
    get:
      parameters:
      - description: Prefix
        in: query
        name: q
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardSuccessEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
      summary: Suggest hashtags (autocomplete)
      tags:
      - Hashtags
  /hashtags/trending:
    get:
      parameters:
      - description: Max tags
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardSuccessEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
      summary: Trending hashtags
      tags:
      - Hashtags
...
  /packages:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardSuccessEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
      summary: List subscription packages
      tags:
      - Packages
  /packages/{code}:
    get:
      parameters:
      - description: Package code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardSuccessEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
      summary: Get package by code
      tags:
      - Packages
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"